MAINT: Moving read_data.py to the helper directory
Moving the read_data.py module to the helper directory for better understanding. Renamed the module to read_config.py as data is ambiguous and doesn't tell you what the data is.
