+++ cloud-chatops/src/dev.py
@@ -5,7 +5,7 @@ import asyncio
 import argparse
 from argparse import Namespace
 from helper.errors import NoTestCase
-from read_data import validate_required_files, get_config
+from helper.read_config import validate_required_files, get_config
 from events import run_global_reminder, run_personal_reminder
 
 logging.basicConfig(level=logging.DEBUG)
