+++ cloud-chatops/src/events.py
@@ -11,7 +11,7 @@ import schedule
 from helper.data import User
 from slack_reminder_api.pr_reminder import PRReminder
 from find_pr_api.github import FindPRs
-from read_data import get_config, get_token
+from helper.read_config import get_config, get_token
 
 
 def run_global_reminder(channel: str) -> None:
