+++ cloud-chatops/src/find_pr_api/github.py
@@ -2,7 +2,7 @@
 
 from typing import List, Dict
 import requests
-from read_data import get_token
+from helper.read_config import get_token
 from helper.data import PR
