+++ cloud-chatops/src/slack_reminder_api/pr_reminder.py
@@ -3,7 +3,7 @@
 from typing import List
 from slack_sdk import WebClient
 from helper.data import PR, Message
-from read_data import get_config
+from helper.read_config import get_config
 
 
 class PRReminder:
